+++ packages/@headlessui-react/src/components/popover/popover.tsx
@@ -17,6 +17,7 @@ import React, {
   MouseEvent as ReactMouseEvent,
   MutableRefObject,
   Ref,
+  MouseEventHandler,
 } from 'react'
 
 import { Props } from '../../types'
@@ -43,6 +44,8 @@ import { Hidden, Features as HiddenFeatures } from '../../internal/hidden'
 import { useEvent } from '../../hooks/use-event'
 import { useTabDirection, Direction as TabDirection } from '../../hooks/use-tab-direction'
 import { microTask } from '../../utils/micro-task'
+
+type MouseEvent<T> = Parameters<MouseEventHandler<T>>[0]
 
 enum PopoverStates {
   Open,
@@ -128,7 +131,9 @@ function usePopoverContext(component: string) {
 }
 
 let PopoverAPIContext = createContext<{
-  close(focusableElement?: HTMLElement | MutableRefObject<HTMLElement | null>): void
+  close(
+    focusableElement?: HTMLElement | MutableRefObject<HTMLElement | null> | MouseEvent<HTMLElement>
+  ): void
   isPortalled: boolean
 } | null>(null)
 PopoverAPIContext.displayName = 'PopoverAPIContext'
@@ -176,7 +181,9 @@ function stateReducer(state: StateDefinition, action: Actions) {
 let DEFAULT_POPOVER_TAG = 'div' as const
 interface PopoverRenderPropArg {
   open: boolean
-  close(focusableElement?: HTMLElement | MutableRefObject<HTMLElement | null>): void
+  close(
+    focusableElement?: HTMLElement | MutableRefObject<HTMLElement | null> | MouseEvent<HTMLElement>
+  ): void
 }
 
 let PopoverRoot = forwardRefWithAs(function Popover<
@@ -271,19 +278,27 @@ let PopoverRoot = forwardRefWithAs(function Popover<
     popoverState === PopoverStates.Open
   )
 
-  let close = useEvent((focusableElement?: HTMLElement | MutableRefObject<HTMLElement | null>) => {
-    dispatch({ type: ActionTypes.ClosePopover })
+  let close = useEvent(
+    (
+      focusableElement?:
+        | HTMLElement
+        | MutableRefObject<HTMLElement | null>
+        | MouseEvent<HTMLElement>
+    ) => {
+      dispatch({ type: ActionTypes.ClosePopover })
 
-    let restoreElement = (() => {
-      if (!focusableElement) return button
-      if (focusableElement instanceof HTMLElement) return focusableElement
-      if (focusableElement.current instanceof HTMLElement) return focusableElement.current
+      let restoreElement = (() => {
+        if (!focusableElement) return button
+        if (focusableElement instanceof HTMLElement) return focusableElement
+        if ('current' in focusableElement && focusableElement.current instanceof HTMLElement)
+          return focusableElement.current
 
-      return button
-    })()
+        return button
+      })()
 
-    restoreElement?.focus()
-  })
+      restoreElement?.focus()
+    }
+  )
 
   let api = useMemo<ContextType<typeof PopoverAPIContext>>(
     () => ({ close, isPortalled }),